+++ NEWS.md
@@ -1,7 +1,12 @@
-# exampleRPackage 0.2.0
+# macrogrow 0.1
 
-* Changed name from exampleDataPackage to exampleRPackage
-* Added a `NEWS.md` file to track changes to the package.
+* Added `check_grow()` function to check that all inputs for `grow_macroalgae()` are configured correctly
+* Added salinity as mandatory data. May change to optional data in future.
+* Working on some diagnostic functions following the template of `T_range()`
+
+# macrogrow 0.0
+
+* Initial setup with all essential functions